pendulum_receiver1.ino: Replace conversion raw -> degrees -> radians by raw -> radians
For some unknown reason, the code was altered such that the raw rotary encoder values are first converted into degrees and then this value gets converted into radians (with a bad approximation of Pi).

Worse than this, a bias in the wrong direction was introduced in the getPoleAngleDeg() method, instead of using the encoderOrigin variable which represents the correct bias.

I now replaced this by the original implementation, which  was correct.

I also changed the bias from -5 to -10 after experimenting with the new pendulum. Further exploration will show if the bias needs to be parameterized such that it can be set for each pendulum individually.